refactor(registry): Extract registry.Transport interface from lorepackage
Move transport abstraction from lorepackage into a dedicated
internal/registry package. Provider→Transport, GitProvider→gitTransport
(unexported), single NewTransport(Config) constructor. Eliminates all
(*GitProvider) type assertions from Registry — version operations now
delegate through the interface. ForceTags moves into registry.Config so
the transport handles it internally.
